requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
django==5.0.6
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extensions
    #   djangorestframework
    #   drf-spectacular
django-cors-headers==4.4.0
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   drf-spectacular
djangorestframework-camel-case==1.4.2
    # via -r requirements.in
drf-spectacular==0.27.2
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.0.0
    # via pytest
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
packaging==24.1
    # via pytest
pluggy==1.5.0
    # via pytest
psycopg2-binary==2.9.9
    # via -r requirements.in
pytest==8.2.2
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.8.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
python-json-logger==2.0.7
    # via -r requirements.in
pyyaml==6.0.2
    # via drf-spectacular
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
sqlparse==0.5.0
    # via django
uritemplate==4.1.1
    # via drf-spectacular
